requirements.txt
altgraph==0.17.4
packaging==24.2
pillow==11.1.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyQt6==6.8.0
PyQt6-Qt6==6.8.1
PyQt6_sip==13.9.1
setuptools==75.8.0
